requirements.txt
rich>=13.0.0
click>=8.1.0
gitpython>=3.1.0
pydantic>=2.0.0
pyyaml>=6.0.0
watchdog>=3.0.0
httpx>=0.25.0
click
gitpython
httpx
pydantic
pyyaml
rich
watchdog
